Create home page into components and transition to next home pages
The home page was hardcoded before, this would make it difficult to add animations
and make changing attributes, such as colour, complex. So the home page is now made up
of components that change and do not change. The ones that change are the ones that will
be animated and be asked to change colour.

TODO: Add animation to everything and fix slider